+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `run_local` option to `BeakerExecutor` for making some steps run locally.
+
 ### Removed
 
 - Removed PyTorch Lightning integration.